Week2-Photo-Tagger/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
transformers==4.44.2
torch==2.2.2
Pillow==10.4.0
python-multipart==0.0.9
pyyaml==6.0.2
